+++ dart/test/sentry_client_test.dart
@@ -406,6 +406,14 @@ void main() {
 
       expect(capturedAttachments.isNotEmpty, true);
     });
+
+    test('should return empty for when transaction is discarded', () async {
+      final client = fixture.getSut(eventProcessor: DropAllEventProcessor());
+      final tr = SentryTransaction(fixture.tracer);
+      final id = await client.captureTransaction(tr);
+
+      expect(id, SentryId.empty());
+    });
   });
 
   group('SentryClient : apply scope to the captured event', () {