+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Unreleased
 
 * Feat: Support Attachment.addToTransactions (#709)
+* Fix: captureTransaction should return emptyId when transaction is discarded (#713)
 
 # 6.3.0-beta.3